server/requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-filter==25.2
django-js-asset==3.1.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.14
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.1.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
